+++ CHANGELOG.md
@@ -66,3 +66,4 @@
 ## [v1.5.0] - 2025-10-18
 ### Added
 - List Top 20 Performers from NSE 200, NSE 500 and BSE 200
+- Added Incomplete Trades with no buy/sell entries